+++ include/utilities/crc32.h
@@ -2,7 +2,7 @@
 
 #include <string>
 
-using uint = unsigned int;
+using size_t = size_t;
 
 namespace nemesis
 {
@@ -12,27 +12,27 @@ namespace nemesis
         CRC32();
         ~CRC32();
 
-        bool FileCRC(const char* sFileName, uint* iOutCRC);
-        bool FileCRC(const char* sFileName, uint* iOutCRC, size_t iBufferSize);
-        uint FileCRC(const char* sFileName);
-        uint FileCRC(const char* sFileName, size_t iBufferSize);
+        bool FileCRC(const char* sFileName, size_t* iOutCRC);
+        bool FileCRC(const char* sFileName, size_t* iOutCRC, size_t iBufferSize);
+        size_t FileCRC(const char* sFileName);
+        size_t FileCRC(const char* sFileName, size_t iBufferSize);
 
-        void FullCRC(const unsigned char* sData, size_t iLength, uint* iOutCRC);
-        void FullCRC(const char* sData, size_t iLength, uint* iOutCRC);
-        void FullCRC(const std::string& sData, uint* iOutCRC);
-        uint FullCRC(const unsigned char* sData, size_t iDataLength);
-        uint FullCRC(const char* sData, size_t iDataLength);
-        uint FullCRC(const std::string& sData);
+        void FullCRC(const unsigned char* sData, size_t iLength, size_t* iOutCRC);
+        void FullCRC(const char* sData, size_t iLength, size_t* iOutCRC);
+        void FullCRC(const std::string& sData, size_t* iOutCRC);
+        size_t FullCRC(const unsigned char* sData, size_t iDataLength);
+        size_t FullCRC(const char* sData, size_t iDataLength);
+        size_t FullCRC(const std::string& sData);
 
-        void PartialCRC(uint* iCRC, const unsigned char* sData, size_t iDataLength);
+        void PartialCRC(size_t* iCRC, const unsigned char* sData, size_t iDataLength);
 
     private:
-        uint Reflect(uint iReflect, const char cChar);
+        size_t Reflect(size_t iReflect, const char cChar);
 
         //0x04C11DB7 is the official polynomial used by PKZip, WinZip and Ethernet.
-        uint iPolynomial = 0x04C11DB7;
-        uint initial     = 0;
-        uint finalxor    = 0;
-        uint iTable[256];
+        size_t iPolynomial = 0x04C11DB7;
+        size_t initial     = 0;
+        size_t finalxor    = 0;
+        size_t iTable[256];
     };
 }
